// zdzw/zdzw-mini-api-gateway-ui/src/app/models/route-filter.interface.ts
/*
 * SPDX-FileCopyrightText: Copyright 2025 Software GmbH
 * SPDX-License-Identifier: Apache-2.0
 */
export enum RouteFilterType {
    ZDZW_LICENSE_VERIFICATION = "ZDZW_LICENSE_VERIFICATION",
    ZDZW_SMARTCONTRACT_VERIFICATION = "ZDZW_SMARTCONTRACT_VERIFICATION",
    EXTERNAL_CALL = "EXTERNAL_CALL",
    SET_HEADERS = "SET_HEADERS",
    OAUTH2_CLIENT = "OAUTH2_CLIENT",
    REMOVE_REQUEST_HEADERS = "REMOVE_REQUEST_HEADERS"
}

export interface RouteFilter {
    type: RouteFilterType;
}

export enum ExternalCall_HTTPMethod {
    CONNECT = "CONNECT",
    DELETE = "DELETE",
    GET = "GET",
    HEAD = "HEAD",
    OPTIONS = "OPTIONS",
    POST = "POST",
    PUT = "PUT",
    PATCH = "PATCH",
    TRACE = "TRACE"
}

export interface ZDZWLicenseVerification_RouteFilter extends RouteFilter {
    type: RouteFilterType.ZDZW_LICENSE_VERIFICATION;
    appId: string;
}

export interface ZDZWSmartContractVerification_RouteFilter extends RouteFilter {
    type: RouteFilterType.ZDZW_SMARTCONTRACT_VERIFICATION;
}

export interface ExternalCall_RouteFilter extends RouteFilter {
    type: RouteFilterType.EXTERNAL_CALL;
    uri: string;
    method: ExternalCall_HTTPMethod | string;
    body?: string;
    headers?: {[key: string]: string | string[]};
    responseMapping?: {[key: string]: string};
}

export interface OAuth2Client_RouteFilter extends RouteFilter {
    type: RouteFilterType.OAUTH2_CLIENT;
    clientId: string;
    clientSecret: string;
    issuerUrl: string;
}

export interface SetHeaders_RouteFilter extends RouteFilter {
    type: RouteFilterType.SET_HEADERS;
    headers: {[key: string]: string};
}

export interface RemoveRequestHeaders_RouteFilter extends RouteFilter {
    type: RouteFilterType.REMOVE_REQUEST_HEADERS;
    headersToRemove: string[];
}